Put all BASIS C++ utilities in sbia::basis namespace. Do not use sbia::<project>::basis namespace.
Even though each project will have their own version of the BASIS C++ utilities, are these only meant for use in .cxx files and not in public header files. Further, each .cxx file may only make use of the BASIS C++ utilities of the own project, not the ones of another BASIS project. Therefore, the name conflicts are indeed desired to prevent this. The use of the BASIS C++ utilities becomes more intuitive when all BASIS C++ utilities, if project-specific or build as part of BASIS, are in the same namespace.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1446 3c61e6fb-ae43-4039-8cf7-c895b3257837