First look for optional dependencies then for required ones. In any case, look for Slicer first.
By looking for optional dependencies first, it is possible that a project can specify that it would be able to use a newer version if available, but otherwise only requires an older version. For example, a project can specify that it could work with ITK 4, but requires at least ITK 3.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1550 3c61e6fb-ae43-4039-8cf7-c895b3257837