Add CMake Find* module for BLAS.
This module has been copied from the distribution of CMake 2.8.5 and patched such that it finds the ATLAS installation on the cluster of SBIA, where instead of the C function cblas_dgemm_, the function cblas_dgemm is compiled into the found static library. Note that missing leading trailing underscore (_) character of the function name.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1767 3c61e6fb-ae43-4039-8cf7-c895b3257837