Do not redirect STDERR of subprocesses to their STDOUT.
Instead, only STDOUT can be captured using the execute_process() functions. Error messages of the subprocess always are printed to the parent's STDERR stream after the subprocess is terminated. This is only a temporary solution. The ideal implementation would make use of non-blocking reads on the subprocess pipes and redirect also STDERR directly when the error message is produced by the child.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1887 3c61e6fb-ae43-4039-8cf7-c895b3257837