Add USE_<Pkg> only if package is not searched for quietly.
The QUIET option is usually used to hide from the user that we were looking for a certain package as it might not be of major importance, respectively, only to identify what is available and what not. BASIS itself contains basis_find_package() calls for the supported interpreters, for example.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1979 3c61e6fb-ae43-4039-8cf7-c895b3257837