Do not set by default <Pkg>_INCLUDE_DIRS and <Pkg>_LIBRARY_DIRS to all directories imported from dependencies.
This seemed to have introduced some issue already in case of BASIS. Assume that the basistest-driver executable of BASIS is statically linked to ITK. In this case, a package which depends on BASIS does not need to use the same ITK version as the one BASIS was built with. If now the BASISConfig.cmake file adds the paths to the header files of that ITK version to its BASIS_INCLUDE_DIRS variable, the dependent package *must* however use the same ITK version. Therefore, set these variables by default only to the paths of the package itself and let package developers add additional paths if required in their ConfigSettings.cmake file.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/branches/basis-1.2@2049 3c61e6fb-ae43-4039-8cf7-c895b3257837