Implement print_wrapped() function and rename term.h/.cxx to stdio.h/.cxx.
The print_wrapped() function is inspired by the TCLAP::StdOutput::spacePrint() implementation of the TCLAP library written by Michael E. Smoot. It further fixes a bug in this implementation which does no proper indentation of text which contains itself newline characters if the text seems to fit on one line.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2180 3c61e6fb-ae43-4039-8cf7-c895b3257837