Revise build of script targets. Revise TargetTools.cmake and MatlabTools.cmake functions. Document target properties.
This is a major change including a change of the behavior of the basis_add_executable() and basis_add_library() functions. Moreover, certain options of these functions or their respective sub-functions were removed. In particular, the WITH_PATH and WITH_EXT options are no longer required. Instead, a new build target SCRIPT_LIBRARY has been added which builds a set of script modules.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2493 3c61e6fb-ae43-4039-8cf7-c895b3257837