Add NAMESPACE argument to basis_project() which allows to use a namespace identifier different from the project name.
Meant to be used in cases where the BASIS_NAMESPACE is part of the project name, but should not be repeated in the namespace. For example, the project name SBIAUtilities results in a PROJECT_NAMESPACE_CMAKE of "sbia.sbiautilities" by default. Setting PROJECT_NAMESPACE to "Utilities", it becomes "sbia.utilities" instead.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2570 3c61e6fb-ae43-4039-8cf7-c895b3257837