Fix bug in _basis_executabletargetinfo_sanitize() which manifests itself when there is a file/directory named "s", "p", "a" "c", "e", "l", "m", "n", "u" in the current working directory.
If there exists a file or directory named "a", the pattern "[:a:]" evaluates to "a". Indeed, it always evaluates to "a" if only one "a" letter appears between the colons.

$ echo [:a:]
a
$ echo [:bcade:]
a
$ echo [:b:]
[:b:]

The same is true for any file/directory whose one character-wide name matches any of the characters in between the colons.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2968 3c61e6fb-ae43-4039-8cf7-c895b3257837