Refactor project attribute CMake variables, project logos, and BASIS configuration.
Use shorter names such as PROJECT_DEVISION_NAME instead of PROJECT_PROVIDER_DEVISION_NAME. Further, to be consistent with the other variables, add _NAME suffix to PROJECT_PACKAGE, PROJECT_PROVIDER, and PROJECT_DEVISION.

During the CMake configuration of BASIS, the default values of most organization specific variables can be set. These are then stored in the BASISConfig.cmake file and used as defaults if a project does not override them.