Issue #254 Rearrange sidebar items on web page once more.
With the new order, using the "next" button in the top right corner should lead the user through the documentation in the recommended sequence. The detailed download and installation pages are at the bottom as these are not needed if one follows the First Steps given in the Quick Start section.