#354 Enable project user and developer to disable the INSTALL_RPATH handling of BASIS.
A project developer can override the CMAKE_SKIP_RPATH setting in config/Settings.cmake. When set to TRUE/ON, BASIS will not set the INSTALL_RPATH property of executables and shared libraries. The option BASIS_INSTALL_RPATH can be used during the project configuration to have BASIS to configure the INSTALL_RPATH. If OFF, BASIS does nothing and lets CMake set this property.