#352 Call basis_add_module only if PROJECT_IS_MODULE.
The PROJECT_MODULES, PROJECT_MODULES_ENABLED, and PROJECT_MODULES_DISABLED variables are in fact not reset for each module but can be read by all modules of the top-level project. This is needed to identify dependencies of a module which are in fact modules of the same top-level project. See basis_find_package.